requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
numpy==1.24.3
boto3==1.34.0
kafka-python==2.0.2